[ADD] (#34) - add slider tooltip
