+++ happyfarmer/package.json
@@ -9,15 +9,16 @@
     "@testing-library/user-event": "^13.5.0",
     "apexcharts": "^3.35.3",
     "axios": "^0.27.2",
-    "react-slick": "^0.29.0",
-    "slick-carousel": "^1.8.1",
-    "rc-slider": "^10.0.0",
+    "rc-slider": "^8.6.6",
+    "rc-tooltip": "^3.7.3",
     "react": "^18.2.0",
     "react-apexcharts": "^1.4.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "react-select": "^5.4.0",
+    "react-slick": "^0.29.0",
+    "slick-carousel": "^1.8.1",
     "styled-components": "^5.3.5",
     "web-vitals": "^2.1.4"
   },
